pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hireplanner"
version = "0.1.0"
description = "Weekly forecasting and labor planning for 3PL warehouses"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "lightgbm>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[project.scripts]
hireplanner = "hireplanner.pipeline.runner:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that require model downloads (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 100
